Fixed terrain light priorities (fix for #139)
 * lights illuminating the terrain specified in scene file are now
   always moved to front of light ordering

+++ src/graphics/opengl/gldevice.cpp
@@ -92,7 +92,7 @@ void CGLDevice::DebugLights()
     m_worldMat.LoadIdentity();
     UpdateModelviewMatrix();
 
-    for (int i = 0; i < m_lights.size(); ++i)
+    for (int i = 0; i < static_cast<int>( m_lights.size() ); ++i)
     {
         color.h = static_cast<float>(i) / static_cast<float>(m_lights.size());
         if (m_lightsEnabled[i])
